requirements.txt
certifi 
cffi==1.17.1
charset-normalizer==3.4.1 
cryptography==44.0.2 
deprecated==1.2.18 
GPUtil==1.4.0 
idna==3.10 
keyboard==0.13.5 
psutil==7.0.0 
pycparser==2.22 
PyGithub==2.6.1 
PyJWT==2.10.1 
PyNaCl==1.5.0 
pyperclip==1.9.0 
PyQt5==5.15.11 
PyQt5-Qt5==5.15.2 
PyQt5_sip==12.17.0 
python-dotenv==1.1.0 
pywin32==310 
requests==2.32.3 
typing_extensions==4.13.2 
urllib3==2.4.0 
wrapt==1.17.2